fix(applying): refactor common utils
