+++ pallas-applying/src/lib.rs
@@ -3,14 +3,14 @@
 pub mod alonzo;
 pub mod byron;
 pub mod shelley_ma;
-pub mod types;
+pub mod utils;
 
 use alonzo::validate_alonzo_tx;
 use byron::validate_byron_tx;
 use pallas_traverse::{Era, MultiEraTx};
 use shelley_ma::validate_shelley_ma_tx;
 
-pub use types::{
+pub use utils::{
     Environment, MultiEraProtParams, UTxOs, ValidationError::TxAndProtParamsDiffer,
     ValidationResult,
 };
