+++ pallas-applying/src/utils/environment.rs
